Working with Possible PMID Matches
This script is going to take a raw file exported from neo4j, get it into the format of a list of PMIDs, then use PubMed API to retrieve full text articles for all those PMIDs. After that, I will use that full text data to check if those articles really do contain the disease that is a supposed "match" for that PMID... Right now I am able to get access the PMIDs from the API but need to work on a specific url to get just the full text in a readable format.